# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * FinanceApp API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type TransactionType =
  (typeof TransactionType)[keyof typeof TransactionType];

export const TransactionType = {
  income: "income",
  expense: "expense",
} as const;

export interface Transaction {
  id: number;
  type: TransactionType;
  amount: number;
  description: string;
  origin?: string | null;
  date: string;
  account?: string | null;
  paymentMethod?: string | null;
  categoryId?: number | null;
  categoryName?: string | null;
  categoryIcon?: string | null;
  categoryColor?: string | null;
  recurring: boolean;
  recurringFrequency?: string | null;
  notes?: string | null;
  createdAt: string;
}

export interface TransactionListResponse {
  transactions: Transaction[];
  total: number;
  page: number;
  limit: number;
  totalIncome: number;
  totalExpenses: number;
  balance: number;
}

export type CreateTransactionBodyType =
  (typeof CreateTransactionBodyType)[keyof typeof CreateTransactionBodyType];

export const CreateTransactionBodyType = {
  income: "income",
  expense: "expense",
} as const;

export interface CreateTransactionBody {
  type: CreateTransactionBodyType;
  amount: number;
  description: string;
  origin?: string | null;
  date: string;
  account?: string | null;
  paymentMethod?: string | null;
  categoryId?: number | null;
  recurring: boolean;
  recurringFrequency?: string | null;
  notes?: string | null;
}

export interface Category {
  id: number;
  name: string;
  icon: string;
  color: string;
  isDefault: boolean;
  transactionCount: number;
  totalAmount: number;
}

export interface CreateCategoryBody {
  name: string;
  icon: string;
  color: string;
}

export interface DashboardSummary {
  balance: number;
  totalIncome: number;
  totalExpenses: number;
  month: string;
  withinLimit: boolean;
}

export interface CategoryBreakdown {
  categoryId: number;
  categoryName: string;
  categoryIcon: string;
  categoryColor: string;
  total: number;
  percentage: number;
  count: number;
}

export type ListTransactionsParams = {
  type?: ListTransactionsType;
  categoryId?: number;
  search?: string;
  period?: ListTransactionsPeriod;
  page?: number;
  limit?: number;
};

export type ListTransactionsType =
  (typeof ListTransactionsType)[keyof typeof ListTransactionsType];

export const ListTransactionsType = {
  income: "income",
  expense: "expense",
} as const;

export type ListTransactionsPeriod =
  (typeof ListTransactionsPeriod)[keyof typeof ListTransactionsPeriod];

export const ListTransactionsPeriod = {
  this_month: "this_month",
  last_month: "last_month",
  last_90_days: "last_90_days",
  all: "all",
} as const;