# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: FinanceApp API specification
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: transactions
    description: Transaction operations
  - name: categories
    description: Category operations
  - name: dashboard
    description: Dashboard summary operations
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /transactions:
    get:
      operationId: listTransactions
      tags: [transactions]
      summary: List all transactions
      parameters:
        - name: type
          in: query
          schema:
            type: string
            enum: [income, expense]
        - name: categoryId
          in: query
          schema:
            type: integer
        - name: search
          in: query
          schema:
            type: string
        - name: period
          in: query
          schema:
            type: string
            enum: [this_month, last_month, last_90_days, all]
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 10
      responses:
        "200":
          description: List of transactions
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TransactionListResponse"
    post:
      operationId: createTransaction
      tags: [transactions]
      summary: Create a transaction
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateTransactionBody"
      responses:
        "201":
          description: Created transaction
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Transaction"

  /transactions/{id}:
    get:
      operationId: getTransaction
      tags: [transactions]
      summary: Get transaction by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Transaction
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Transaction"
    put:
      operationId: updateTransaction
      tags: [transactions]
      summary: Update a transaction
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateTransactionBody"
      responses:
        "200":
          description: Updated transaction
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Transaction"
    delete:
      operationId: deleteTransaction
      tags: [transactions]
      summary: Delete a transaction
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  /categories:
    get:
      operationId: listCategories
      tags: [categories]
      summary: List all categories
      responses:
        "200":
          description: List of categories
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Category"
    post:
      operationId: createCategory
      tags: [categories]
      summary: Create a custom category
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateCategoryBody"
      responses:
        "201":
          description: Created category
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Category"

  /categories/{id}:
    put:
      operationId: updateCategory
      tags: [categories]
      summary: Update a category
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateCategoryBody"
      responses:
        "200":
          description: Updated category
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Category"
    delete:
      operationId: deleteCategory
      tags: [categories]
      summary: Delete a custom category
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  /dashboard/summary:
    get:
      operationId: getDashboardSummary
      tags: [dashboard]
      summary: Get dashboard summary with balance, income, expenses
      responses:
        "200":
          description: Dashboard summary
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DashboardSummary"

  /dashboard/recent-transactions:
    get:
      operationId: getRecentTransactions
      tags: [dashboard]
      summary: Get most recent transactions for dashboard
      responses:
        "200":
          description: Recent transactions
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Transaction"

  /dashboard/category-breakdown:
    get:
      operationId: getCategoryBreakdown
      tags: [dashboard]
      summary: Get spending breakdown by category
      responses:
        "200":
          description: Category breakdown
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/CategoryBreakdown"

components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status

    Transaction:
      type: object
      properties:
        id:
          type: integer
        type:
          type: string
          enum: [income, expense]
        amount:
          type: number
        description:
          type: string
        origin:
          type: string
          nullable: true
        date:
          type: string
          format: date
        account:
          type: string
          nullable: true
        paymentMethod:
          type: string
          nullable: true
        categoryId:
          type: integer
          nullable: true
        categoryName:
          type: string
          nullable: true
        categoryIcon:
          type: string
          nullable: true
        categoryColor:
          type: string
          nullable: true
        recurring:
          type: boolean
        recurringFrequency:
          type: string
          nullable: true
        notes:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - type
        - amount
        - description
        - date
        - recurring
        - createdAt

    TransactionListResponse:
      type: object
      properties:
        transactions:
          type: array
          items:
            $ref: "#/components/schemas/Transaction"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
        totalIncome:
          type: number
        totalExpenses:
          type: number
        balance:
          type: number
      required:
        - transactions
        - total
        - page
        - limit
        - totalIncome
        - totalExpenses
        - balance

    CreateTransactionBody:
      type: object
      properties:
        type:
          type: string
          enum: [income, expense]
        amount:
          type: number
        description:
          type: string
        origin:
          type: string
          nullable: true
        date:
          type: string
          format: date
        account:
          type: string
          nullable: true
        paymentMethod:
          type: string
          nullable: true
        categoryId:
          type: integer
          nullable: true
        recurring:
          type: boolean
        recurringFrequency:
          type: string
          nullable: true
        notes:
          type: string
          nullable: true
      required:
        - type
        - amount
        - description
        - date
        - recurring

    Category:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        icon:
          type: string
        color:
          type: string
        isDefault:
          type: boolean
        transactionCount:
          type: integer
        totalAmount:
          type: number
      required:
        - id
        - name
        - icon
        - color
        - isDefault
        - transactionCount
        - totalAmount

    CreateCategoryBody:
      type: object
      properties:
        name:
          type: string
        icon:
          type: string
        color:
          type: string
      required:
        - name
        - icon
        - color

    DashboardSummary:
      type: object
      properties:
        balance:
          type: number
        totalIncome:
          type: number
        totalExpenses:
          type: number
        month:
          type: string
        withinLimit:
          type: boolean
      required:
        - balance
        - totalIncome
        - totalExpenses
        - month
        - withinLimit

    CategoryBreakdown:
      type: object
      properties:
        categoryId:
          type: integer
        categoryName:
          type: string
        categoryIcon:
          type: string
        categoryColor:
          type: string
        total:
          type: number
        percentage:
          type: number
        count:
          type: integer
      required:
        - categoryId
        - categoryName
        - categoryIcon
        - categoryColor
        - total
        - percentage
        - count