# lib/api-zod/src/generated/api.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * FinanceApp API specification
 * OpenAPI spec version: 0.1.0
 */
import * as zod from "zod";

/**
 * @summary Health check
 */
export const HealthCheckResponse = zod.object({
  status: zod.string(),
});

/**
 * @summary List all transactions
 */
export const listTransactionsQueryPageDefault = 1;
export const listTransactionsQueryLimitDefault = 10;

export const ListTransactionsQueryParams = zod.object({
  type: zod.enum(["income", "expense"]).optional(),
  categoryId: zod.coerce.number().optional(),
  search: zod.coerce.string().optional(),
  period: zod
    .enum(["this_month", "last_month", "last_90_days", "all"])
    .optional(),
  page: zod.coerce.number().default(listTransactionsQueryPageDefault),
  limit: zod.coerce.number().default(listTransactionsQueryLimitDefault),
});

export const ListTransactionsResponse = zod.object({
  transactions: zod.array(
    zod.object({
      id: zod.number(),
      type: zod.enum(["income", "expense"]),
      amount: zod.number(),
      description: zod.string(),
      origin: zod.string().nullish(),
      date: zod.coerce.date(),
      account: zod.string().nullish(),
      paymentMethod: zod.string().nullish(),
      categoryId: zod.number().nullish(),
      categoryName: zod.string().nullish(),
      categoryIcon: zod.string().nullish(),
      categoryColor: zod.string().nullish(),
      recurring: zod.boolean(),
      recurringFrequency: zod.string().nullish(),
      notes: zod.string().nullish(),
      createdAt: zod.coerce.date(),
    }),
  ),
  total: zod.number(),
  page: zod.number(),
  limit: zod.number(),
  totalIncome: zod.number(),
  totalExpenses: zod.number(),
  balance: zod.number(),
});

/**
 * @summary Create a transaction
 */
export const CreateTransactionBody = zod.object({
  type: zod.enum(["income", "expense"]),
  amount: zod.number(),
  description: zod.string(),
  origin: zod.string().nullish(),
  date: zod.coerce.date(),
  account: zod.string().nullish(),
  paymentMethod: zod.string().nullish(),
  categoryId: zod.number().nullish(),
  recurring: zod.boolean(),
  recurringFrequency: zod.string().nullish(),
  notes: zod.string().nullish(),
});

/**
 * @summary Get transaction by ID
 */
export const GetTransactionParams = zod.object({
  id: zod.coerce.number(),
});

export const GetTransactionResponse = zod.object({
  id: zod.number(),
  type: zod.enum(["income", "expense"]),
  amount: zod.number(),
  description: zod.string(),
  origin: zod.string().nullish(),
  date: zod.coerce.date(),
  account: zod.string().nullish(),
  paymentMethod: zod.string().nullish(),
  categoryId: zod.number().nullish(),
  categoryName: zod.string().nullish(),
  categoryIcon: zod.string().nullish(),
  categoryColor: zod.string().nullish(),
  recurring: zod.boolean(),
  recurringFrequency: zod.string().nullish(),
  notes: zod.string().nullish(),
  createdAt: zod.coerce.date(),
});

/**
 * @summary Update a transaction
 */
export const UpdateTransactionParams = zod.object({
  id: zod.coerce.number(),
});

export const UpdateTransactionBody = zod.object({
  type: zod.enum(["income", "expense"]),
  amount: zod.number(),
  description: zod.string(),
  origin: zod.string().nullish(),
  date: zod.coerce.date(),
  account: zod.string().nullish(),
  paymentMethod: zod.string().nullish(),
  categoryId: zod.number().nullish(),
  recurring: zod.boolean(),
  recurringFrequency: zod.string().nullish(),
  notes: zod.string().nullish(),
});

export const UpdateTransactionResponse = zod.object({
  id: zod.number(),
  type: zod.enum(["income", "expense"]),
  amount: zod.number(),
  description: zod.string(),
  origin: zod.string().nullish(),
  date: zod.coerce.date(),
  account: zod.string().nullish(),
  paymentMethod: zod.string().nullish(),
  categoryId: zod.number().nullish(),
  categoryName: zod.string().nullish(),
  categoryIcon: zod.string().nullish(),
  categoryColor: zod.string().nullish(),
  recurring: zod.boolean(),
  recurringFrequency: zod.string().nullish(),
  notes: zod.string().nullish(),
  createdAt: zod.coerce.date(),
});

/**
 * @summary Delete a transaction
 */
export const DeleteTransactionParams = zod.object({
  id: zod.coerce.number(),
});

/**
 * @summary List all categories
 */
export const ListCategoriesResponseItem = zod.object({
  id: zod.number(),
  name: zod.string(),
  icon: zod.string(),
  color: zod.string(),
  isDefault: zod.boolean(),
  transactionCount: zod.number(),
  totalAmount: zod.number(),
});
export const ListCategoriesResponse = zod.array(ListCategoriesResponseItem);

/**
 * @summary Create a custom category
 */
export const CreateCategoryBody = zod.object({
  name: zod.string(),
  icon: zod.string(),
  color: zod.string(),
});

/**
 * @summary Update a category
 */
export const UpdateCategoryParams = zod.object({
  id: zod.coerce.number(),
});

export const UpdateCategoryBody = zod.object({
  name: zod.string(),
  icon: zod.string(),
  color: zod.string(),
});

export const UpdateCategoryResponse = zod.object({
  id: zod.number(),
  name: zod.string(),
  icon: zod.string(),
  color: zod.string(),
  isDefault: zod.boolean(),
  transactionCount: zod.number(),
  totalAmount: zod.number(),
});

/**
 * @summary Delete a custom category
 */
export const DeleteCategoryParams = zod.object({
  id: zod.coerce.number(),
});

/**
 * @summary Get dashboard summary with balance, income, expenses
 */
export const GetDashboardSummaryResponse = zod.object({
  balance: zod.number(),
  totalIncome: zod.number(),
  totalExpenses: zod.number(),
  month: zod.string(),
  withinLimit: zod.boolean(),
});

/**
 * @summary Get most recent transactions for dashboard
 */
export const GetRecentTransactionsResponseItem = zod.object({
  id: zod.number(),
  type: zod.enum(["income", "expense"]),
  amount: zod.number(),
  description: zod.string(),
  origin: zod.string().nullish(),
  date: zod.coerce.date(),
  account: zod.string().nullish(),
  paymentMethod: zod.string().nullish(),
  categoryId: zod.number().nullish(),
  categoryName: zod.string().nullish(),
  categoryIcon: zod.string().nullish(),
  categoryColor: zod.string().nullish(),
  recurring: zod.boolean(),
  recurringFrequency: zod.string().nullish(),
  notes: zod.string().nullish(),
  createdAt: zod.coerce.date(),
});
export const GetRecentTransactionsResponse = zod.array(
  GetRecentTransactionsResponseItem,
);

/**
 * @summary Get spending breakdown by category
 */
export const GetCategoryBreakdownResponseItem = zod.object({
  categoryId: zod.number(),
  categoryName: zod.string(),
  categoryIcon: zod.string(),
  categoryColor: zod.string(),
  total: zod.number(),
  percentage: zod.number(),
  count: zod.number(),
});
export const GetCategoryBreakdownResponse = zod.array(
  GetCategoryBreakdownResponseItem,
);