# lib/api-zod/src/generated/types/category.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * FinanceApp API specification
 * OpenAPI spec version: 0.1.0
 */

export interface Category {
  id: number;
  name: string;
  icon: string;
  color: string;
  isDefault: boolean;
  transactionCount: number;
  totalAmount: number;
}