# lib/api-zod/src/generated/types/createCategoryBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * FinanceApp API specification
 * OpenAPI spec version: 0.1.0
 */

export interface CreateCategoryBody {
  name: string;
  icon: string;
  color: string;
}