# lib/api-zod/src/generated/types/createTransactionBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * FinanceApp API specification
 * OpenAPI spec version: 0.1.0
 */
import type { CreateTransactionBodyType } from "./createTransactionBodyType";

export interface CreateTransactionBody {
  type: CreateTransactionBodyType;
  amount: number;
  description: string;
  origin?: string | null;
  date: Date;
  account?: string | null;
  paymentMethod?: string | null;
  categoryId?: number | null;
  recurring: boolean;
  recurringFrequency?: string | null;
  notes?: string | null;
}