# lib/api-zod/src/generated/types/dashboardSummary.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * FinanceApp API specification
 * OpenAPI spec version: 0.1.0
 */

export interface DashboardSummary {
  balance: number;
  totalIncome: number;
  totalExpenses: number;
  month: string;
  withinLimit: boolean;
}