# lib/api-zod/src/generated/types/transaction.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * FinanceApp API specification
 * OpenAPI spec version: 0.1.0
 */
import type { TransactionType } from "./transactionType";

export interface Transaction {
  id: number;
  type: TransactionType;
  amount: number;
  description: string;
  origin?: string | null;
  date: Date;
  account?: string | null;
  paymentMethod?: string | null;
  categoryId?: number | null;
  categoryName?: string | null;
  categoryIcon?: string | null;
  categoryColor?: string | null;
  recurring: boolean;
  recurringFrequency?: string | null;
  notes?: string | null;
  createdAt: Date;
}