# lib/api-zod/src/generated/types/transactionListResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * FinanceApp API specification
 * OpenAPI spec version: 0.1.0
 */
import type { Transaction } from "./transaction";

export interface TransactionListResponse {
  transactions: Transaction[];
  total: number;
  page: number;
  limit: number;
  totalIncome: number;
  totalExpenses: number;
  balance: number;
}